+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     - `CampionamentoHFBI::check_record`
   - Type changes:
     - `domain::hfbi::RIFERIMENTO_HFBI`: `LazyLock<Vec<SpecieHFBI>>` -> `LazyLock<RiferimentoHFBI>`
-
+- Drop `'static` bound on `csv::deser` methods
 - Drop implicit logging on check functions
   - Usercode can use `csv::stanis::giorgio::format_csv_errors` explicitly instead
 - Drop internal transitional methods